requirements.txt
torch
transformers
Pillow
pandas
requests
